+++ code/spikes/dev.000/package.json
@@ -25,8 +25,8 @@
   "private": true,
   "types": "./dist/types.d/index.html.d.html",
   "exports": {
-    ".": "./dist/web/lib/index-d549cb01.js",
-    "./web": "./dist/web/lib/index-d549cb01.js"
+    ".": "./dist/web/lib/index-fefffd77.js",
+    "./web": "./dist/web/lib/index-fefffd77.js"
   },
   "typesVersions": {
     "*": {
