+++ code/spikes/dev.000/src/test.ui/entry.Specs.mts
@@ -1,4 +1,4 @@
-import { SelfDevSpecs as DevSpecs } from 'sys.ui.react.dev';
+import { ModuleSpecs as DevSpecs } from 'sys.ui.react.dev';
 import { Specs as CommonSpecs } from 'sys.ui.react.common';
 import { Specs as MonacoSpecs } from 'sys.ui.react.monaco';
 import { Specs as VideoSpecs } from 'sys.ui.react.video';
