+++ code/system.ui/react.dev/src/index.mts
@@ -1,6 +1,5 @@
 export { Pkg } from './index.pkg.mjs';
 
 export { Dev } from './Dev.mjs';
-export { Spec } from './Spec';
-
-export { SelfDevSpecs as SelfDevSpecs } from './test.ui/entry.Specs.mjs';
+export { Spec } from './logic.Spec';
+export { ModuleSpecs } from './test.ui/entry.Specs.mjs';
