+++ code/system.ui/react.dev/src/logic.Bus/common.ts
@@ -1,3 +1,3 @@
 export * from '../common';
-export { Spec } from '../Spec';
+export { Spec } from '../logic.Spec';
 export { DEFAULT } from '../DEFAULT.mjs';
