+++ code/system.ui/react.dev/src/test.ui/common.ts
@@ -1,3 +1,3 @@
 export * from '../common';
-export { Spec } from '../Spec';
+export { Spec } from '../logic.Spec';
 export { TestLog } from '../test.ui';
