+++ code/system.ui/react.dev/src/test.ui/entry.Specs.mts
@@ -1,8 +1,8 @@
+export const ModuleSpecs = {
+  'sys.ui.dev.Harness': () => import('../ui/Harness/Harness.SPEC'),
+};
+
 export const SampleSpecs = {
   'sample.MySample': () => import('../test.ui/sample.specs/MySample.SPEC'),
   'sample.DevTools.Button': () => import('../test.ui/sample.DevTools/s.ui.Button.SPEC'),
-};
-
-export const SelfDevSpecs = {
-  'sys.ui.dev.Harness': () => import('../ui/Harness/Harness.SPEC'),
 };
