+++ code/system.ui/react.dev/src/test.ui/index.ts
@@ -1,4 +1,4 @@
-export { Spec } from '../Spec';
+export { Spec } from '../logic.Spec';
 export { Dev } from '../Dev.mjs';
 
 export * from '../common';
