+++ code/system.ui/react.dev/src/ui/common.ts
@@ -1,4 +1,5 @@
 export * from '../common';
 export * from '../logic.Bus';
 export * from '../ui.Hooks';
-export { Spec } from '../Spec';
+
+export { Spec } from '../logic.Spec';
